TextEncoder API Extensions
I'm curious whether it's as feasible to just use the base `TextEncoder` and `TextDecoder` APIs, and modify their inputs and outputs so it's MUTF-8 compatible, rather than rebuilding the UTF-8 parsing and writing from scratch for the whole thing. If this does work, is fast, and passes all of the tests in this project, I may likely take that route for NBTify, since I'm already using the default standard lib text encoders, so I don't know if it really makes sense to fully reinvent the wheel just for this. Depends on the amount of changes necessary to ensure this is compatible I guess. We'll see! Just curious if this is a route I should take. Maybe this can be upstreamed back to this library if it is worthwhile to the maintainer as well? It's not my first goal, since I don't know how they feel about this kind of setup, but might as well explore it since it's an idea :P

These changes still pass the Jest tests, that's one of the main reasons I'm trialing this out in this fork, since I want to borrow the tests to see if this version will work.