AssemblyScript Demo
Things aren't working according to the tests, I'm going to stick with using the plain JS implementation itself. This was a learning-packed demo to set this up though! I borrowed aspects of this from my WASM-AS project. The AssemblyScript docs have a lot of nice information on how to set things up accordingly, and what is supported in either JS or WASM.

https://www.assemblyscript.org/status.html
https://www.assemblyscript.org/compiler.html

https://github.com/sciencesakura/mutf-8/issues/23